requirements.txt
# python3 -m  pipreqs.pipreqs . --force
# use  pip show Jinja2 to check package version
python-dotenv==0.21.0

# for scraping 
requests==2.28.1
beautifulsoup4==4.11.1
lxml==4.9.1

# for sending email
PyPDF2==2.10.6
redmail==0.4.1      